Minor fixes.
* src/getopt.m4 (print_option_descr): Fix type of the width argument.
* src/rcfile.c (esmtp_kw): Remove unised variable.
* src/tls.c (start_ssl_server): Fix type of the status variable/
* src/anubisusr.c: Fix typecasts.
* src/daemon.c: Likewise.
* src/logport.c: Likewise.
* src/net.c: Likewise.
* src/stream.c: Likewise.

+++ src/daemon.c
@@ -313,7 +313,7 @@ static int
 _stdio_write (void *sd, const char *data, size_t size, size_t * nbytes)
 {
   int rc;
-  int fd = (int) sd;
+  int fd = (int) (ptrdiff_t) sd;
 
   if (fd == 0)
     fd = 1;
@@ -330,7 +330,7 @@ static int
 _stdio_read (void *sd, char *data, size_t size, size_t * nbytes)
 {
   int n;
-  int fd = (int) sd;
+  int fd = (int) (ptrdiff_t) sd;
   fd_set rds;
 
   errno = 0;
